refactor(frontend): relocate project files and update `pnpm-lock.yaml`
Several project files related to dependencies have been moved around which is reflected in updated `pnpm-lock.yaml` file. Furthermore, `pnpm-lock.yaml` has been updated with new dependencies such as '@types/node', 'typescript' and 'vite', and large amounts of resolution data. The relocation of files refines the project's structure and the new dependencies will add to the functionality of the project.